Summation
For this challenge, I had to find the summation of every number between 1 and num. For this case, i = 0.

As long as 0 is less than the number inputted, I will add 1.

If the number inputted is greater than one, I will add the number inputted and the number subtracted by one.
Otherwise, I will just return 1.