Vowel Counting
I have to return the number of vowels in the given string. For this case, just A E I O U is fine. The string will only consist of lower case letters and/or spaces.

With my inputted strings, I set the number of vowels to 0.
For the input, if anything entered matched AEIOU (lower or uppercase), I would add that instance to the num_vowels variable.  I would then return the number of vowels given into string.