Hex Dump Program
This program is a hex dumping tool for inspecting binary files!
The first argument made was the most necessary, which was "file".
The With command, in this instance, was used twice. 
Once to open the file to read in one to dump the results out.
My current hangup is trying to get the (ord) function out.
The tutorial I used mentioned this but it's.. redundant, so I am working on modifying that!